backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.6
stripe==7.6.0
python-dotenv==1.0.0
pydantic==2.4.2
aiosqlite==0.19.0
python-dateutil==2.8.2
openai==1.0.0
